# restapi/swaggerui/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "this is to show how to write RESTful APIs in golang.\nthat is to provide a detailed overview of the language specs",
    "title": "Account API.",
    "contact": {
      "name": "Jordan Rhodes",
      "email": "mydocs@example.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "paths": {
    "/accounts/": {
      "post": {
        "description": "If account creation is success, account will be returned with Created (201).",
        "tags": [
          "accounts"
        ],
        "summary": "Creates a new account.",
        "operationId": "createAccount",
        "parameters": [
          {
            "description": "account to add to the list of accounts",
            "name": "account",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Account"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/okResp"
          },
          "400": {
            "$ref": "#/responses/badReq"
          }
        }
      }
    },
    "/accounts/{id}": {
      "get": {
        "description": "If the account is found, account will be returned else Error Not Found (404) will be returned.",
        "tags": [
          "accounts"
        ],
        "summary": "Return an Account provided by the id.",
        "operationId": "getAccount",
        "parameters": [
          {
            "type": "string",
            "description": "id of the account",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/accountRes"
          },
          "400": {
            "$ref": "#/responses/badReq"
          },
          "404": {
            "$ref": "#/responses/notFoundReq"
          }
        }
      },
      "delete": {
        "description": "Depending on the account id, HTTP Status Not Found (404) or HTTP Status OK (200) may be returned.",
        "tags": [
          "accounts"
        ],
        "summary": "Deletes requested account by account id.",
        "operationId": "deleteAccount",
        "parameters": [
          {
            "type": "string",
            "description": "account id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/okResp"
          },
          "400": {
            "$ref": "#/responses/badReq"
          },
          "404": {
            "$ref": "#/responses/notFoundReq"
          }
        }
      }
    }
  },
  "definitions": {
    "Account": {
      "description": "Account request model",
      "type": "object",
      "properties": {
        "first_name": {
          "description": "First name of the account holder",
          "type": "string",
          "x-go-name": "FirstName"
        },
        "id": {
          "description": "Id of the account",
          "type": "string",
          "x-go-name": "ID"
        },
        "last_name": {
          "description": "Last name of the account holder",
          "type": "string",
          "x-go-name": "LastName"
        },
        "user_name": {
          "description": "Username of the account holder",
          "type": "string",
          "x-go-name": "UserName"
        }
      },
      "x-go-package": "_/Users/jordanrhodes/Developer/playground/learning-go/restapi"
    }
  },
  "responses": {
    "accountRes": {
      "description": "Account response payload",
      "schema": {
        "$ref": "#/definitions/Account"
      }
    },
    "badReq": {
      "description": "Error Bad Request",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code 400 -  Bad Request",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Code"
          }
        }
      }
    },
    "notFoundReq": {
      "description": "Error Not Found",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code 404 -  Not Found",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Code"
          }
        }
      }
    },
    "okResp": {
      "description": "Success response",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code 200 - OK",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Code"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "KEY",
      "in": "header"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}